requirements.txt
pytest==8.0.0
requests==2.32.3
allure-pytest==2.13.5
urllib3==2.3.0
python-dotenv~=1.0.1
pydantic==2.10.6
assertpy==1.1
playwright==1.36.0
pyparsing==3.1.0
pytest-base-url==2.0.0
pytest-playwright==0.3.3